+++ tests/testdata/check_lua/pattern_many_matches.py
@@ -1,0 +1,6 @@
+data = [
+    # <block check-lua="tests/testdata/check_lua/echo.lua" check-lua-pattern="id: (?P<value>\d+)" name="42,7">
+    name: Alice, id: 42
+    name: Bob, id: 7
+    # </block>
+]